Add functionality to the Reservation screen
In this issue, we added functionality to the FlatList component to be
able to view the flights stored in the database. We achieved this by
querying Firebase and specifying that we wanted the data related to
the current user's ID. We used the useEffect hook to ensure that the
query is only executed once and not on every render.

Co-authored-by: Tole15 <rodrigermoon@gmail.com>
Co-authored-by: BrandonVG <brandon.vadillox3@gmail.com>